BoardShowページのcustom hooksを書いた
