+++ lib/client/selectors/board.ts
@@ -1,2 +1,4 @@
-export const findBoard = (boards, boardId) =>
+import { Board } from '../../../model/client/Bard'
+
+export const findBoard = (boards: Board[], boardId): Board | undefined =>
   boards.find((board) => board.id === boardId)
